♻️ [Refactoring]: deployment::bash_escape モジュール分離
- src/component/deployment/bash_escape.rs に write_executable_script / escape_for_bash_double_quote を抽出（pub(super)）
- src/component/deployment/bash_escape_test.rs に escape 系 8 件 + write 系 3 件 = 11 件を移動
- 挙動不変: 親ディレクトリ未作成時の Err 伝播、改行 \\n エスケープを維持
- 既存 N=59 件 green 維持
